pyproject.toml
[tool.poetry]
name = "repl_python3_AcadBot"
version = "0.1.0"
description = ""
authors = ["Dabidesu <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
replit = "^3.1.0"
pip = "^21.2.4"
google-auth-oauthlib = "^0.4.6"
google-api-python-client = "^2.19.1"
google_api = "^0.1.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
